feat(logs): add logs grouping

+++ app/src/main/java/kite1412/irrigo/model/WateringConfig.kt
@@ -2,6 +2,6 @@ package kite1412.irrigo.model
 
 data class WateringConfig(
     val minSoilMoisturePercent: Float, // 0 - 100%
-    val durationMs: Float,
+    val durationMs: Int,
     val automated: Boolean
 )
